embed test not relying on external data files in the package
